+++ Assets/Adons/Sky Inventory/Scripts/itemName.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public class itemName : MonoBehaviour
 {
-    public int ItemPrise;
+    public float ItemPrise;
     public string ItemData = "";
     public string _Name;
 }
